+++ doc/whats_new.rst
@@ -24,7 +24,7 @@ Version 0.3 (Unreleased)
 Enhancements
 ~~~~~~~~~~~~
 
-- ...
+- Adding symmetric orthogonalization via :func:`mne_connectivity.symmetric_orth`, by `Eric Larson`_ (:gh:`36`)
 
 Bug
 ~~~
@@ -40,6 +40,7 @@ Authors
 ~~~~~~~
 
 * `Adam Li`_
+* `Eric Larson`_
 
 :doc:`Find out what was new in previous releases <whats_new_previous_releases>`